Per issue #128 - Start of code review comments.

+++ BostonFreedomTrail/Utils/Constants.swift
@@ -34,7 +34,7 @@ enum SegueConstants : String {
     case MapToPlacemarkSegueIdentifier = "MapToPlacemarkSegueIdentifier"
 }
 
-enum TabBarControllerTags : NSInteger {
+enum TabBarControllerTags : Int {
     case MapViewTag = 0
     case VirtualTourViewTag = 1
     case AboutViewTag = 2
